% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{as.data.frame.flowFrame}
\alias{as.data.frame.flowFrame}
\title{coerce the flowFrame to a data.frame}
\usage{
\method{as.data.frame}{flowFrame}(x, ...)
}
\arguments{
\item{x}{flowFrame}
}
\value{
data.frame
}
\description{
It extracts the cell event matrix and convert it to a data.frame.
}

