% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{as.data.frame.flowSet}
\alias{as.data.frame.flowSet}
\title{coerce the flowSet to a data.frame}
\usage{
\method{as.data.frame}{flowSet}(x, ...)
}
\arguments{
\item{x}{flowSet}
}
\value{
data.frame
}
\description{
It extract the cell event matrix from each flowFrame
 and combind them to a single data.frame.
}

