% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-flowSet.R
\name{as.filterList.data.frame}
\alias{as.filterList.data.frame}
\title{Convert data.frame back to filterList}
\usage{
as.filterList.data.frame(df, pcols = ".rownames")
}
\arguments{
\item{pcols}{the pData columns}
}
\description{
It is used for gating purporse for geom_stats layer
(no longer needed since the data is now not foritfied until print.ggcyo)
}

