% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-flowSet.R
\name{as.gate.data.frame}
\alias{as.gate.data.frame}
\title{Convert data.frame back to original gate format}
\usage{
as.gate.data.frame(df)
}
\description{
Convert data.frame back to original gate format
}

