% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto.R
\name{as.ggplot}
\alias{as.ggplot}
\title{It fortifies the data, fills some default settings and returns a regular ggplot object.}
\usage{
as.ggplot(x)
}
\arguments{
\item{x}{ggcyto object with the data that has not yet been fortified to data.frame.}
}
\description{
The orginal data format is preserved during the ggcyo constructor because they still need to be used during the plot building process.
}

