% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/axis_inverse_trans.R
\name{axis_x_inverse_trans}
\alias{axis_x_inverse_trans}
\alias{axis_y_inverse_trans}
\title{Display axis in raw scales}
\usage{
axis_x_inverse_trans(...)

axis_y_inverse_trans(...)
}
\arguments{
\item{...}{common continuous scale parameters passed to 'continuous_scale' (not used currently)}

\item{gs}{GatingSet}

\item{channel}{the channel or marker name}
}
\description{
It is essentially a dummy continous scale and will be instantiated
by '+.ggcyto_GatingSet' with 'breaks` and 'lables' customized.
}

