% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/btextGrob.R
\name{btextGrob}
\alias{btextGrob}
\title{Grid Text with a Background (modified based on Kmisc::grid.text2)}
\usage{
btextGrob(label, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  just = "centre", hjust = NULL, vjust = NULL, check.overlap = FALSE,
  default.units = "npc", name = NULL, gp = gpar(col = "black", fill =
  "grey92", lineend = "butt", linejoin = "round"), gp.rect = gpar(col =
  "white", fill = "white", lineend = "butt", linejoin = "round"), vp = NULL,
  widthAdj = unit(0.05, "npc"), heightAdj = unit(0.05, "npc"))
}
\arguments{
\item{label}{A character or \code{\link{expression}} vector.
Other objects are coerced by \code{as.graphicsAnnot}.}

\item{x}{A numeric vector or unit object specifying x-values.}

\item{y}{A numeric vector or unit object specifying y-values.}

\item{just}{The justification of the text relative to its (x, y) location.
If there are two values, the first value specifies horizontal justification
and the second value specifies vertical justification. Possible string values
are: \code{"left"}, \code{"right"}, \code{"centre"}, \code{"center"},
\code{"bottom"}, and \code{"top"}.
For numeric values, 0 means left alignment and 1 means right alignment.}

\item{hjust}{A numeric vector specifying horizontal justification.
If specified, overrides the \code{just} setting.}

\item{vjust}{A numeric vector specifying vertical justification.
If specified, overrides the \code{just} setting.}

\item{check.overlap}{A logical value to indicate whether to check for
and omit overlapping text.}

\item{default.units}{A string indicating the default units to use
if \code{x} or \code{y} are only given as numeric vectors.}

\item{name}{A character identifier.}

\item{gp}{An object of class \code{gpar}, typically the output from a call to the function
\code{gpar}. This is basically a list of graphical parameter settings.}

\item{vp}{A Grid viewport object (or \code{NULL}).}

\item{widthAdj}{A width adjustment parameter, to help control how much horizontal padding
there should be between the text and the background rectangle.}

\item{heightAdj}{A height adjustment parameter, to help control how much
vertical padding there should be between the text and the background rectangle.}

\item{draw}{A logical value indiciating whether graphics output should be produced.}
}
\description{
Grid Text with a Background (modified based on Kmisc::grid.text2)
}
\seealso{
\code{\link{grid.text}} and \code{\link{grid.rect}}
}

