% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_stats.percent}
\alias{compute_stats.percent}
\title{compute the statistics of the gates and its position in the respective projections}
\usage{
compute_stats.percent(fs, gates, digits = 3, ...)
}
\arguments{
\item{fs}{flowSet}

\item{gates}{a list of filters}

\item{digits}{control the percent format}

\item{...}{other arguments passed to stat_position function}
}
\value{
a data.frame that contains percent and centroid locations as well as pData
}
\description{
It is used to produce the data for geom_btext layer.
}

