% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faust_gating_plot.R
\name{faust_gating_plot}
\alias{faust_gating_plot}
\title{plot faust gating schemes}
\usage{
faust_gating_plot(gh, start_node, end_node, ...)
}
\arguments{
\item{start_node}{faust start node}

\item{end_node}{the terminal leaf node generated by faust}
}
\description{
plot faust gating schemes
}
\examples{
\dontrun{
gs=load_gs("~/Downloads/ics")

end_node = "/S/LV/L/CD4+/CD3+/CD8-/TNF+/CD107a-/IL4-/IFNg+/IL2+/CD154-/IL17a-"
start_node = "/S/LV/L"
gh=gs[[1]]
p = faust_gating_plot(gh, start_node, end_node, bins=128)
plot(ggcyto_arrange(p, nrow=1))
}
}
