% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_flowCore_fasinh.R
\name{flow_breaks}
\alias{flow_breaks}
\title{flow_breaks
breaks for flow}
\usage{
flow_breaks(x, n = 6, equal.space = FALSE, trans.fun, inverse.fun)
}
\arguments{
\item{x}{data}

\item{n}{number of breaks}

\item{equal.space}{should it be equal sapced}

\item{trans.fun}{actual transform function}

\item{inverse.fun}{inverse transform function}
}
\description{
flow_breaks
breaks for flow
}
