% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_flowCore_fasinh.R
\name{flow_trans}
\alias{flow_trans}
\title{generate a trans objects
Used by other specific trans constructor}
\usage{
flow_trans(name, trans.fun, inverse.fun, equal.space = FALSE, n = 6)
}
\arguments{
\item{name}{transformation name}
}
\description{
generate a trans objects
Used by other specific trans constructor
}
