% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_biexp.R
\name{flowJo_biexp_breaks}
\alias{flowJo_biexp_breaks}
\title{flowJo biexponential breaks (integer breaks on biexp-transformed scales)}
\usage{
flowJo_biexp_breaks(n = 6, pretty = FALSE, ...)
}
\arguments{
\item{n}{desired number of breaks}

\item{pretty}{if use log10-like breaks that is normally used in flowJo
when FALSE, breaks are equally spaced instead.}

\item{...}{parameters passed to flowWorkspace::flowJoTrans function}
}
\description{
flowJo biexponential breaks (integer breaks on biexp-transformed scales)
}

