% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_biexp.R
\name{flowJo_biexp_trans}
\alias{flowJo_biexp_trans}
\title{flowJo biexponential transformation.}
\usage{
flowJo_biexp_trans(..., pretty = FALSE)
}
\arguments{
\item{...}{arguments passed to flowJoTrans}
}
\description{
flowJo biexponential transformation.
}

