% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_fasinh.R
\name{flowJo_fasinh_breaks}
\alias{flowJo_fasinh_breaks}
\title{flowJo inverse hyperbolic sine breaks (integer breaks on fasinh-transformed scales)}
\usage{
flowJo_fasinh_breaks(n = 6, ...)
}
\arguments{
\item{n}{desired number of breaks}

\item{...}{parameters passed to flowJo.fasinh}
}
\description{
flowJo inverse hyperbolic sine breaks (integer breaks on fasinh-transformed scales)
}

