% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_fasinh.R
\name{flowJo_fasinh_trans}
\alias{flowJo_fasinh_trans}
\title{flowJo inverse hyperbolic sine transformation.}
\usage{
flowJo_fasinh_trans(...)
}
\arguments{
\item{...}{parameters passed to flowJo.fasinh}
}
\value{
fasinh transformation object
}
\description{
Used for inverse hyperbolic sine scale layer \code{\link{scale_x_flowJo_fasinh}}
}
\examples{
flowJo_fasinh_trans()
}

