% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.ellipsoidGate}
\alias{fortify.ellipsoidGate}
\title{Convert a ellipsoidGate to a data.table useful for ggplot}
\usage{
\method{fortify}{ellipsoidGate}(model, data, ...)
}
\arguments{
\item{model}{ellipsoidGate}

\item{data}{not used.}

\item{...}{not used.}
}
\description{
It converts ellipsoidGate to polygongate before fortifying it.
}

