% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.flowFrame}
\alias{fortify.flowFrame}
\title{Convert a flowFrame to a ggplot-compatible data.table}
\usage{
\method{fortify}{flowFrame}(model, data, ...)
}
\arguments{
\item{model}{flowFrame}

\item{data}{not used.}

\item{...}{not used.}
}
\description{
It actually converts the flowFrame to flowSet first and
then dispatch to the fority method for flowSet.
}

