% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.GatingSet}
\alias{fortify.GatingSet}
\title{coerce a GatingSet node to data.table}
\usage{
\method{fortify}{GatingSet}(model, ...)
}
\arguments{
\item{model}{GatingSet}

\item{...}{not used.}
}
\description{
coerce a GatingSet node to data.table
}

