% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.multiRangeGate}
\alias{fortify.multiRangeGate}
\title{Convert a multiRangeGate to a data.table useful for ggplot}
\usage{
\method{fortify}{multiRangeGate}(model, data = NULL, ...)
}
\arguments{
\item{model}{multiRangeGate}

\item{data}{Not used}

\item{...}{not used.}

\item{nPoints}{not used}
}
\value{
data.table
}
\description{
It converts the boundaries slot into a data.table
}
\examples{
mrq = multiRangeGate(ranges = list(min=c(100, 350), max=c(250, 400)))
fortify(mrq)
}
