% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.polygonGate}
\alias{fortify.polygonGate}
\title{Convert a polygonGate to a data.table useful for ggplot}
\usage{
\method{fortify}{polygonGate}(model, data, measure_range = NULL,
  bins = NULL, ...)
}
\arguments{
\item{model}{polygonGate}

\item{data}{not used.}

\item{measure_range,bins}{the plot information collected from flow data and geom_hex used to interpolate polygon with more vertices. Interpolation is mainly
for the purpose of plotting (so that it won't lose its shape from subsetting through 'limits').
But it is not necessary for other purposes like centroid calculation.}

\item{...}{not used.}
}
\value{
data.table
}
\description{
It converts the boundaries slot into a data.table
When 'bins' is supplied, the method tries to interpolate the polygon with more verticies.
The number of verticies increases if larger `bins` is specified. (represents resolution of geom_hex).
}
\examples{
sqrcut <- matrix(c(300,300,600,600,50,300,300,50),ncol=2,nrow=4)
colnames(sqrcut) <- c("FSC-H","SSC-H")
pg <- polygonGate(filterId="nonDebris", .gate= sqrcut)
fortify(pg) #no interpolation
fortify(pg, bins = 30) # with interpolation
}

