% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_fs.R
\name{fortify_fs.flowFrame}
\alias{fortify_fs.flowFrame}
\title{coerce flowFrame to flowSet
The default coerce method does not perserve the sample name.}
\usage{
\method{fortify_fs}{flowFrame}(model, data, ...)
}
\arguments{
\item{model}{flowFrame}

\item{data}{not used.}

\item{...}{not used.}
}
\description{
coerce flowFrame to flowSet
The default coerce method does not perserve the sample name.
}

