% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_fs.R
\name{fortify_fs.GatingSet}
\alias{fortify_fs.GatingSet}
\title{coerce a GatingSet node to flowSet
The default coerce method does not perserve the sample name.}
\usage{
\method{fortify_fs}{GatingSet}(model, data, ...)
}
\arguments{
\item{model}{GatingSet object that has 'subset' character attribute that specifies the node name}
}
\description{
coerce a GatingSet node to flowSet
The default coerce method does not perserve the sample name.
}

