% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gate.R
\name{gate_null}
\alias{gate_null}
\title{clear all the geom_gate() layer previously added}
\usage{
gate_null()
}
\description{
clear all the geom_gate() layer previously added
}
\examples{
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
#autoplot display pop stats by default
p <- autoplot(gs, "CD4")
#it is easy to remove the default gate
p <- p + gate_null()
#and add a new one
p <- p + geom_gate("CD8")
p
}
