% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_btext.R
\name{geom_btext}
\alias{geom_btext}
\title{modified geom_text that draws the rectangle background.}
\usage{
geom_btext(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", parse = FALSE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{stat}{The statistical transformation to use on the data for this
layer.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{parse}{If TRUE, the labels will be parsed into expressions and
displayed as described in ?plotmath}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}
}
\description{
Limitation: It does not support rotation.
}
\examples{
\donttest{
ggplot(mtcars, aes(x=wt, y=mpg, label=rownames(mtcars))) + geom_btext(size = 5, fill = "gray")
}
}

