% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_gate.R
\name{geom_gate.character}
\alias{geom_gate.character}
\title{construct an abstract geom layer for a character that represents nodes in a Gating tree}
\usage{
\method{geom_gate}{character}(data, ...)
}
\description{
It will be instanatiated later as a specific geom_gate layer or layers based on the gates
extracted from the given GatingSet object.
}

