% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_gate.R
\name{geom_gate.list}
\alias{geom_gate.filterList}
\alias{geom_gate.list}
\title{Add layer for flowCore::filterList}
\usage{
\method{geom_gate}{list}(data, ...)

\method{geom_gate}{filterList}(data, pd, ...)
}
\arguments{
\item{data}{a named list}

\item{pd}{pData (data.frame) that has rownames represents the sample names used as key to be merged with filterList}

\item{data}{filterList}
}
\description{
pdata of the flow data must be supplied here explicitly in order for
the gates to be dispatched to each panel.
}

