% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_gate.R
\name{geom_gate.polygonGate}
\alias{geom_gate.polygonGate}
\title{construct geom layer for polygonGate}
\usage{
\method{geom_gate}{polygonGate}(data, mapping = NULL, fill = "transparent",
  colour = "red", ...)
}
\arguments{
\item{data}{a polygonGate}

\item{fill}{polygonGate is not filled by default}

\item{colour}{default is red}

\item{The}{mapping aesthetic mapping}
}
\description{
Not intended to be used directly.
}

