% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_gate.R
\name{geom_gate}
\alias{geom_gate}
\alias{geom_gate.default}
\title{add a flowCore gate layer}
\usage{
geom_gate(data, ...)

\method{geom_gate}{default}(data, ...)
}
\arguments{
\item{data}{a rectangleGate, polygonGate or ellipsoidGate}
}
\description{
Currently only rectangleGate (1d or 2d), polygonGate, ellipsoidGate are supported.
}

