% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_gate.R
\name{geom_gate.rectangleGate}
\alias{geom_gate.rectangleGate}
\title{construct geom layer for rectangleGate}
\usage{
\method{geom_gate}{rectangleGate}(data, mapping = NULL,
  fill = "transparent", colour = "red", ...)
}
\arguments{
\item{fill}{rectangleGate is not filled by default}

\item{colour}{default is red}

\item{The}{mapping aesthetic mapping}

\item{gate}{a rectangleGate}
}
\description{
Not intended to be used directly.
}

