% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_hvline.R
\name{geom_hvline}
\alias{geom_hvline}
\title{Vertical or horizontal line.}
\usage{
geom_hvline(mapping = NULL, data = NULL, position = "identity",
  show_guide = FALSE, ...)
}
\arguments{
\item{show_guide}{should a legend be drawn? (defaults to \code{FALSE})}
}
\description{
This geom is based on the source code of ' \code{\link{geom_hline}} and \code{\link{geom_vline}}.
}
\details{
The goal is to determine the line to be either vertial or horizontal based on the 1-d data provided in
this layer.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{ggplot2:::rd_aesthetics("geom", "vline")}
}
\examples{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
# vline
p + geom_hvline(data = data.frame(wt= 3))
# hline
p + geom_hvline(data = data.frame(mpg= 20))
}
\seealso{
\code{\link{geom_hline}} for horizontal lines,
 \code{\link{geom_segment}} for a more general approach"
}

