% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_stats.R
\name{geom_stats}
\alias{geom_stats}
\title{Popluation statistics layer}
\usage{
geom_stats(..., type = "percent")
}
\arguments{
\item{...}{other arguments passed to stat_position function}

\item{type}{can be "percent", "count" or "MFI".}
}
\description{
It is a dummy layer that simply inherits GeomBText.
The actual data for geom_text layer will be instanatiated within ggycto.+ operatoer
by looking up the geom_gate layer.
So it is dedicated for ggcyto context and thus can't not be added to ggplot object directly.
}
\examples{
p <- ggcyto(fs, aes(x = `FSC-H`, y =  `SSC-H`))
p <- p + geom_hex(bins = 128)
p + geom_gate(rect.gates) + geom_stats()
}

