% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsne.R
\name{geom_tsne_degree}
\alias{geom_tsne_degree}
\alias{geom_tsne_marker}
\alias{geom_tsne_poly}
\title{Some geom_layers to be added to  "ggcyto_tsne" object}
\usage{
geom_tsne_degree(size = 1.5, alpha = 0.5, ...)

geom_tsne_marker(marker, size = 1.5, alpha = 0.5, ...)

geom_tsne_poly(degree = 3, count = 20, size = 1.5, alpha = 0.5, ...)
}
\arguments{
\item{size, alpha, ...}{arguments passed to geom_point}

\item{marker}{one of the specific marker name used in "stat_tsne". It will be used as "colour" aesthetic for geom_tsne_marker layer.}

\item{degree}{a threshold that sets the minimum degrees of functionality to be displayed in geom_tsne_poly layer}

\item{count}{a threshold that sets the minimum cell count of polyfunctionality to be displayed in geom_tsne_poly layer}
}
\description{
They are the abstract layers that will be instantiated as the "geom_point" 
by overloaded "+" operator when they are added to \link{ggcyto_tsne} object.
with some customized settings to visualize the "Degree of Functionality", "polyfunctional" or "marker signal".
}
\details{
See examples in \link{stat_tsne}.
}

