% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-GatingSet.R
\name{getChildren_by_projection}
\alias{getChildren_by_projection}
\title{match the subpopulation based on the given projections and parentID}
\usage{
getChildren_by_projection(gs, parentID, x, y)
}
\description{
match the subpopulation based on the given projections and parentID
}

