\name{ggCyto-package}
\alias{ggCyto-package}
\alias{ggCyto}
\docType{package}
\title{
\packageTitle{ggCyto}
}
\description{
\packageDescription{ggCyto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ggCyto}
\packageIndices{ggCyto}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ggCyto}

Maintainer: \packageMaintainer{ggCyto}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
