% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto.R
\name{ggcyto.flowSet}
\alias{ggcyto.flowSet}
\title{Create a new ggcyto plot from a flowSet}
\usage{
\method{ggcyto}{flowSet}(data, mapping, ...)
}
\arguments{
\item{data}{default flowSet for plot}

\item{mapping}{default list of aesthetic mappings (these can be colour,
size, shape, line type -- see individual geom functions for more details)}

\item{...}{ignored}

\item{environment}{in which evaluation of aesthetics should occur}
}
\description{
Create a new ggcyto plot from a flowSet
}

