% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-gs.R
\name{ggcyto.GatingSet}
\alias{ggcyto.GatingSet}
\title{Create a new ggcyto plot from a flowSet}
\usage{
\method{ggcyto}{GatingSet}(data, parent, mapping, ...)
}
\arguments{
\item{data}{GatingSet to plot}

\item{mapping}{default list of aesthetic mappings (these can be colour,
size, shape, line type -- see individual geom functions for more details)}

\item{...}{ignored}
}
\description{
Create a new ggcyto plot from a flowSet
}

