% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-GatingSet.R
\docType{methods}
\name{+.ggcyto_GatingSet}
\alias{+,ggcyto_GatingSet,ANY-method}
\alias{+.ggcyto_GatingSet}
\title{overloaded '+' method for ggcyto.gs}
\usage{
\method{+}{ggcyto_GatingSet}(e1, e2)

\S4method{+}{ggcyto_GatingSet,ANY}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{ggcyto}}

\item{e2}{A component to add to \code{e1}}
}
\description{
It tries to copy pData from ggcyto object to the gate layers
so that the gate layer does not need to have `pd` to be supplied explicitly by users.
}

