% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto_par.R
\name{ggcyto_par_default}
\alias{ggcyto_par_default}
\title{Return The default ggcyto settings}
\usage{
ggcyto_par_default()
}
\value{
a list of default settings for ggycto
}
\description{
Return The default ggcyto settings
}
\examples{
ggcyto_par_default()
}

