% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcyto_tsne.R
\docType{methods}
\name{+.ggcyto_tsne}
\alias{+,ggcyto_tsne,ANY-method}
\alias{+.ggcyto_tsne}
\title{overloaded '+' method for ggcyto_tsne}
\usage{
\method{+}{ggcyto_tsne}(e1, e2)

\S4method{+}{ggcyto_tsne,ANY}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{ggcyto_tsne}}

\item{e2}{A component to add to \code{e1}}
}
\value{
ggcyto_tsne object
}
\description{
It tries to parse geom_tsne_* layers added ggcyto_tsne object.
}

