% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_data_frame.R
\name{ggplot.data.frame}
\alias{ggplot.data.frame}
\title{copied from ggplot2 2.2.1 since it was removed from 2.3}
\usage{
\method{ggplot}{data.frame}(data, mapping = aes(), ...,
  environment = parent.frame())
}
\description{
copied from ggplot2 2.2.1 since it was removed from 2.3
}
