% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-flowSet.R
\name{is.geom_gate_filterList}
\alias{is.geom_gate_filterList}
\title{Checking if a layer is geom_gate layer for a filterList
by checking If the layer data is annotated.}
\usage{
is.geom_gate_filterList(layer)
}
\description{
TODO: It will be more robust to define a new type of proto object to for
this type of idenitiy checking.
}

