% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto.R
\name{is.ggcyto}
\alias{is.ggcyto}
\title{Reports whether x is a ggcyto object}
\usage{
is.ggcyto(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE/FALSE
}
\description{
Reports whether x is a ggcyto object
}
\examples{
data(GvHD)
fs <- GvHD[1:2]
p <- ggcyto(fs, aes(x = `FSC-H`))
is.ggcyto(p)
}

