% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-flowSet.R
\name{is.ggcyto_flowSet}
\alias{is.ggcyto_flowSet}
\title{Reports whether x is a ggplot object}
\usage{
is.ggcyto_flowSet(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a ggplot object
}

