% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theme.R
\name{is.ggcyto_par}
\alias{is.ggcyto_par}
\title{Reports whether x is a ggcyto_par object}
\usage{
is.ggcyto_par(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a ggcyto_par object
}

