% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theme.R
\name{is.ggcyto_theme}
\alias{is.ggcyto_theme}
\title{Reports whether x is a ggcyto_theme object}
\usage{
is.ggcyto_theme(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a ggcyto_theme object
}

