% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcyto_tsne.R
\name{is.ggcyto_tsne}
\alias{is.ggcyto_tsne}
\title{Reports whether x is a ggcyto_tsne object}
\usage{
is.ggcyto_tsne(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE or FALSE
}
\description{
Reports whether x is a ggcyto_tsne object
}

