% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_logicle.R
\name{logicle_breaks}
\alias{logicle_breaks}
\title{flowCore logicle breaks (integer breaks on logicle-transformed scales)}
\usage{
logicle_breaks(n = 6, ...)
}
\arguments{
\item{n}{desired number of breaks}

\item{...}{parameters passed to flowJo.fasinh}
}
\value{
a function that generates logicle space
}
\description{
Used to construct \code{\link{logicle_trans}} object
}
\examples{
data <- 1:1e3
brks.func <- logicle_breaks()
brks <- brks.func(data)
brks # logicle space displayed at raw data scale

#logicle transform it to verify it is equal-spaced at transformed scale
trans.obj <- logicleTransform()
trans.func <- slot(trans.obj, ".Data")
brks.trans <- trans.func(brks)
brks.trans
}

