% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_logicle.R
\name{logicle_trans}
\alias{logicle_trans}
\title{logicle transformation.}
\usage{
logicle_trans(...)
}
\arguments{
\item{...}{arguments passed to logicleTransform.}
}
\value{
a logicle transformation object
}
\description{
Used for logicle scale layer \code{\link{scale_x_logicle}}
}
\examples{
logicle_trans()
}

