% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility.R
\name{marginalFilter}
\alias{marginalFilter}
\title{Generate a marginal gate.}
\usage{
marginalFilter(fs, dims, tol = 1e-05, ...)
}
\arguments{
\item{fs}{flowSet}

\item{dims}{the channels involved}

\item{tol}{the tolerance}

\item{...}{not used}
}
\value{
an expressionFilter
}
\description{
It constructs an expression filter that removes the marginal events.
}

