% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_hvline.R
\name{my_compute_aesthetics}
\alias{my_compute_aesthetics}
\title{override the original Layer$compute_aesthetics method to make the 1d gate data work with geom_segment}
\usage{
my_compute_aesthetics(., data, plot)
}
\description{
This is the hook invoked by ggplot_build at early stage of plotting.
We need to match the 1d gate definition with the mapping
stored in the original aes (from 'plot$mapping') so that the line can be determined as either horizontal or vetical
}
\details{
And so far this is the best place where we can look up the original mapping and store the axis information.
Idealy it could have been be done at `map_statistic` step. Unfortunetly, the data would have already been generalized at that point and
thus it won't be able to do the matching otherwise.
}

