% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoplot.R
\name{+.ggcyto_gate_layout}
\alias{+.ggcyto_gate_layout}
\title{overloaded '+' method for ggcyto_gate_layout}
\usage{
\method{+}{ggcyto_gate_layout}(e1, e2)
}
\arguments{
\item{e1}{ggcyto_gate_layout}

\item{e2}{any ggplot layer}
}
\value{
a modified ggcyto_gate_layout object
}
\description{
It adds the layer specified by 'e2' to each individual ggplot object stored in ggcyto_gate_layout
}

