% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-GatingLayout.R
\docType{methods}
\name{+.ggcyto_GatingLayout}
\alias{+,ggcyto_GatingLayout,ANY-method}
\alias{+.ggcyto_GatingLayout}
\title{overloaded '+' method for ggcyto_gate_layout}
\usage{
\method{+}{ggcyto_GatingLayout}(e1, e2)

\S4method{+}{ggcyto_GatingLayout,ANY}(e1, e2)
}
\arguments{
\item{e1}{ggcyto_gate_layout}

\item{e2}{any ggplot layer}
}
\value{
a modified ggcyto_gate_layout object
}
\description{
It adds the layer specified by 'e2' to each individual ggplot object stored in ggcyto_gate_layout
}

