% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoplot.R, R/ggcyto.R
\docType{methods}
\name{show.ggcyto_gate_layout}
\alias{plot.ggcyto}
\alias{print,ggcyto-method}
\alias{print.ggcyto}
\alias{show,ggcyto-method}
\alias{show,ggcyto_gate_layout-method}
\alias{show.ggcyto}
\alias{show.ggcyto_gate_layout}
\title{Draw ggcyto on current graphics device.}
\usage{
\method{show}{ggcyto_gate_layout}(object)

\S4method{show}{ggcyto_gate_layout}(object)

\method{print}{ggcyto}(x, ...)

\S4method{print}{ggcyto}(x, ...)

\method{plot}{ggcyto}(x, ...)

\method{show}{ggcyto}(object)

\S4method{show}{ggcyto}(object)
}
\arguments{
\item{x}{ggcyto object to display}

\item{...}{other arguments not used by this method}
}
\description{
A wrapper for print.ggplot.
}

