% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoplot.R
\name{print.ggcyto_gate_layout}
\alias{print.ggcyto_gate_layout}
\title{print method for ggcyto_gate_layout class
It calls arrangeGrob to arrange a list of ggplot objects stored as ggcyto_gate_layout object}
\usage{
print.ggcyto_gate_layout(x, ...)
}
\arguments{
\item{x}{ggcyto_gate_layout, which is essentially a list of ggplot objects}

\item{...}{not used}
}
\description{
print method for ggcyto_gate_layout class
It calls arrangeGrob to arrange a list of ggplot objects stored as ggcyto_gate_layout object
}

