% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcyto_GatingLayout.R
\name{print.ggcyto_GatingLayout}
\alias{print.ggcyto_GatingLayout}
\alias{show.ggcyto_GatingLayout}
\alias{show,ggcyto_GatingLayout-method}
\title{print method for ggcyto_gate_layout class}
\usage{
\method{print}{ggcyto_GatingLayout}(x, ...)

\method{show}{ggcyto_GatingLayout}(object)
}
\arguments{
\item{x}{ggcyto_gate_layout, which is essentially a list of ggplot objects that were previously stored as ggcyto_gate_layout object by autoplot function.}

\item{...}{other arguments passed to arrangeGrob}

\item{object}{ggcyto_GatingLayout}
}
\value{
nothing
}
\description{
print method for ggcyto_gate_layout class
}
