% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_logicle.R
\name{scale_x_logicle}
\alias{scale_x_logicle}
\alias{scale_y_logicle}
\title{flowJo inverse hyperbolic sine scale}
\usage{
scale_x_logicle(..., w = 0.5, t = 262144, m = 4.5, a = 0)

scale_y_logicle(..., w = 0.5, t = 262144, m = 4.5, a = 0)
}
\arguments{
\item{...}{common continuous scale parameters passed to 'continuous_scale' (not used currently)}

\item{w,t,m,a}{see 'help(logicleTransform')}
}
\description{
flowJo inverse hyperbolic sine scale
}

