% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_raw.R
\name{scale_x_raw}
\alias{scale_x_raw}
\alias{scale_y_raw}
\title{Display axis in raw scales}
\usage{
scale_x_raw(...)

scale_y_raw(...)
}
\arguments{
\item{...}{common continuous scale parameters passed to 'continuous_scale' (not used currently)}

\item{gs}{GatingSet}

\item{channel}{the channel or marker name}
}
\description{
It is essentially a dummy continous scale and will be instantiated
by '+.ggcyto_GatingSet' with 'breaks` and 'lables' customized.
}

