% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stat_position.R
\name{stat_position}
\alias{stat_position}
\alias{stat_position.filter}
\alias{stat_position.filterList}
\alias{stat_position.list}
\title{compute the positions of the population statistics based on the geometic gate centroid}
\usage{
stat_position(gate, ...)

\method{stat_position}{filter}(gate, adjust = 0.5, abs = FALSE,
  data_range = NULL)

\method{stat_position}{filterList}(gates, ...)

\method{stat_position}{list}(gates, ...)
}
\arguments{
\item{gate}{a flowCore filter}

\item{adjust}{adjust the position of the centroid}

\item{abs}{logical}

\item{data_range}{the actual data range}
}
\value{
the gate centroid coordinates
}
\description{
compute the positions of the population statistics based on the geometic gate centroid
}

