% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tsne.R
\name{stat_tsne}
\alias{stat_tsne}
\alias{stat_tsne.character}
\alias{stat_tsne.default}
\title{tsne layer.}
\usage{
stat_tsne(data, ...)

\method{stat_tsne}{default}(data, ...)

\method{stat_tsne}{character}(data = "_parent_", nodes, groupBy = "name",
  nEvents = NULL, tsne_args = list(check_duplicates = FALSE, theta = 0.9),
  ...)
}
\arguments{
\item{data}{the parent population node to be operated on. If not supplied, it must be specified through 'subset' argument in 'ggcyto' constructor.}

\item{nodes, ...}{the arguments passed to see \link{getSingleCellExpression}}

\item{groupBy}{the phenotype data used to group the samples for subsampling cell events}

\item{nEvents}{the maximum number of cell events to be used for tsne}

\item{tsne_args}{a list of arguments passed to \link{Rtsne}}
}
\description{
This is a virtual layer which only passes on the parameters to ggcyto object without doing any actual computing on itself. 
Once it is added, the ggcyto object is transformed to the 'ggcyto_tsne' class.
The save the time, the tsne calculation is only invoked at the inital plotting action. 
The subsequent plots will use the cached data stored in ggcyto_tsne object.
}
\examples{
\dontrun{
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
p <- ggcyto(gs, subset = "CD8")
p <- p + stat_tsne(nodes = c("CD8/38+ DR-", "CD8/CCR7- 45RA+"), marginal = FALSE, nEvents = 1e3) 
myscale <- scale_color_gradientn(colours = rev(RColorBrewer::brewer.pal(11, "Spectral")))
p + geom_point(aes(color = poly)) # subsequent plotting does not need to recompute tsne
p <- p + myscale
p + geom_point(aes(color = degree)) #plotting first time could be slow due to the inital computing of tsne
p + geom_point(aes(color = `CD38 APC`), size = 1.5, alpha = 0.5)  
}
}

