% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stats.R
\name{stats_null}
\alias{stats_null}
\title{clear all the geom_stats() layer previously added}
\usage{
stats_null()
}
\description{
clear all the geom_stats() layer previously added
}
\examples{
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
#autoplot display pop stats by default
p <- autoplot(gs1, "CD4")
#it is easy to remove the default stats
p <- p + stats_null()
#and add a new one
p <- p + geom_stats(type = "count")
}
