% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ggcyto}
\alias{theme_ggcyto}
\title{Set ggcyto theme elements}
\usage{
theme_ggcyto(...)
}
\arguments{
\item{...}{a list of element name, element pairings that modify the
  existing theme.}
}
\description{
Use this function to modify theme settings.
There are not really theme elements but the regular (or to be instantiated as) scales, labs, facet objects.
They are just organized as a cluster of settings that can be added as a single layer to the plot for the convenience.
}
\section{Theme elements}{

The individual theme elements are:

\tabular{ll}{
  limits             \tab can be "data"(default) or "instrument" or a list of numeric limits for x and y
                   (e.g. \code{list(x = c(0, 4000))}) \cr
  facet             \tab the regular facet object \cr
  hex_fill             \tab default scale_fill_gradientn for geom_hex layer \cr
  lab              \tab labs_cyto object

}
}
\examples{
\dontrun{
p <- ggcyto(gs, aes(x = CD4, y = CD8), subset = "3+")
# 2d plot
p <- p + geom_hex(bins = 64)
p

#use instrument range by overwritting limits setting in the default theme
p + theme_ggcyto(limits = "instrument")

#manually set limits
myTheme <- theme_ggcyto(limits = list(x = c(0,3.2e3), y = c(-10, 3.5e3)))
 p  + myTheme# or xlim(0,3.2e3) + ylim(-10, 3.5e3)
}
}

