% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.flowSet}
\alias{autoplot.GatingHierarchy}
\alias{autoplot.GatingSet}
\alias{autoplot.GatingSetList}
\alias{autoplot.flowFrame}
\alias{autoplot.flowSet}
\alias{autoplot.ncdfFlowList}
\title{Plot fluorescence intensity in one or two dimension.}
\usage{
\method{autoplot}{flowSet}(object, x, y = NULL, bins = 30, ...)

\method{autoplot}{ncdfFlowList}(object, ...)

\method{autoplot}{flowFrame}(object, ...)

\method{autoplot}{GatingSetList}(object, ...)

\method{autoplot}{GatingSet}(object, gate, x = NULL, y = "SSC-A",
  bins = 30, ...)

\method{autoplot}{GatingHierarchy}(object, gate, y = "SSC-A", bool = FALSE,
  arrange.main = sampleNames(object), arrange = TRUE, merge = TRUE,
  projections = list(), strip.text = c("parent", "gate"), ...)
}
\arguments{
\item{object}{flowFrame, flowSet, GatingSet object}

\item{x,y}{define the dimension of the plot}

\item{bins}{passed to geom_hex}

\item{...}{other arguments passed to ggplot}

\item{gate}{the gate to be plotted}

\item{bool}{whether to plot boolean gates}

\item{arrange.main}{the main title of the arranged plots}

\item{arrange}{whether to use arrangeGrob to put multiple plots in the same page}

\item{merge}{wehther to merge multiple gates into the same panel when they share the same parent and projections}

\item{projections}{a list of customized projections}

\item{strip.text}{either "parent" (the parent population name) or "gate "(the gate name). The latter usually is used when merge is FALSE}
}
\value{
a ggcyto object
}
\description{
Overloaded autoplot for the cytomertry data structure: flowFrame or flowSet, Gatinghierarchy, GatingSet.
It plots the cytometry data with geom_histogram, geom_density or geom_hex.
}
\examples{
library(flowCore)
data(GvHD)
fs <- GvHD[subset(pData(GvHD), Patient \%in\%5:7 & Visit \%in\% c(5:6))[["name"]]]

#1d- density plot
autoplot(fs, x = "SSC-H")

#2d plot: default geom_hex plot
autoplot(fs, x = 'FSC-H', y ='SSC-H')

#autplot for GatingSet
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
autoplot(gs, "CD3+")

#autplot for GatingHierarchy
gh <- gs[[1]]
autoplot(gh) # by default the strip.text shows the parent population

#To display the gate name
#autoplot(gh , strip.text = "gate")
}

