% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.flowSet}
\alias{autoplot.GatingHierarchy}
\alias{autoplot.GatingSet}
\alias{autoplot.flowFrame}
\alias{autoplot.flowSet}
\title{Plot fluorescence intensity in one or two dimension.}
\usage{
\method{autoplot}{flowSet}(object, x, y = NULL, bins = 30, ...)

\method{autoplot}{flowFrame}(object, ...)

\method{autoplot}{GatingSet}(object, gate, x = NULL, y = "SSC", ...)

\method{autoplot}{GatingHierarchy}(object, gate, y = "SSC", bool = FALSE,
  arrange.main = sampleNames(object), arrange = TRUE, merge = TRUE,
  projections = list(), ...)
}
\arguments{
\item{object}{flowFrame, flowSet, GatingSet object}

\item{x,y}{define the dimension of the plot}

\item{bins}{passed to geom_hex}

\item{...}{other arguments passed to ggplot}

\item{gate}{the gate to be plotted}

\item{gate}{the gate to be plotted}
}
\value{
a ggcyto object
}
\description{
Overloaded autoplot for the cytomertry data structure: flowFrame or flowSet, Gatinghierarchy, GatingSet.
It plots the cytometry data with geom_histogram, geom_density or geom_hex.
}
\examples{
data(GvHD)
fs <- GvHD[subset(pData(GvHD), Patient \%in\%5:7 & Visit \%in\% c(5:6))[["name"]]]

#1d- density plot
autoplot(fs, x = "SSC-H")

#2d plot: default geom_hex plot
autoplot(fs, x = 'FSC-H', y ='SSC-H')
}

