% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_stats}
\alias{compute_stats}
\title{compute the statistics of the cell population defined by gates}
\usage{
compute_stats(fs = NULL, gates, type = "percent", value = NULL,
  data_range = NULL, ...)
}
\arguments{
\item{fs}{flowSet. can be NULL when precaculated 'value' is provided}

\item{gates}{a list of filters}

\item{type}{can be "percent", "count" or "MFI".}

\item{value}{the pre-calculated stats value. when supplied, the stats computing is skipped.}

\item{data_range}{the data range for each channels}

\item{...}{other arguments passed to stat_position function}
}
\value{
a data.frame that contains percent and centroid locations as well as pData
that used as data for geom_btext layer.
}
\description{
It calls the underlining stats routine and merge it with the lable position calculated by stat_position
as well as the pData of flowSet.
}

