% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_fasinh.R
\name{flowJo.fasinh}
\alias{flowJo.fasinh}
\alias{flowJo.fsinh}
\title{hyperbolic sine/inverse hyperbolic sine (flowJo-version) transform function constructor}
\usage{
flowJo.fasinh(m = 4, t = 12000, a = 0.7, length = 256)

flowJo.fsinh(m = 4, t = 12000, a = 0.7, length = 256)
}
\arguments{
\item{m}{numeric the full width of the transformed display in asymptotic decades}

\item{t}{numeric the maximum value of input data}

\item{a}{numeric Additional negative range to be included in the display in asymptotic decades}

\item{length}{numeric the maximum value of transformed data}
}
\description{
hyperbolic sine/inverse hyperbolic sine (flowJo-version) transform function constructor
}

