% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_biexp.R
\name{flowJo_biexp_breaks}
\alias{flowJo_biexp_breaks}
\title{flowJo biexponential breaks (integer breaks on biexp-transformed scales)}
\usage{
flowJo_biexp_breaks(n = 6, pretty = FALSE, ...)
}
\arguments{
\item{n}{desired number of breaks}

\item{pretty}{if use log10-like breaks that is normally used in flowJo
when FALSE, breaks are equally spaced instead.}

\item{...}{parameters passed to \code{\link[flowWorkspace]{flowJoTrans}}}
}
\value{
a function generates biexponential space
}
\description{
Used to construct \code{\link{flowJo_biexp_trans}} object
}
\examples{
data <- 1:1e3
brks.func <- flowJo_biexp_breaks()
brks <- brks.func(data)
brks # biexp space displayed at raw data scale

#transform it to verify it is equal-spaced at transformed scale
trans.func <- flowJoTrans()
brks.trans <- trans.func(brks)
brks.trans
}

