% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_fasinh.R
\name{flowJo_fasinh_breaks}
\alias{flowJo_fasinh_breaks}
\title{flowJo inverse hyperbolic sine breaks (integer breaks on fasinh-transformed scales)}
\usage{
flowJo_fasinh_breaks(n = 6, ...)
}
\arguments{
\item{n}{desired number of breaks}

\item{...}{parameters passed to flowJo.fasinh}
}
\value{
a function generates fasinh or fsinh space
}
\description{
Used to construct \code{\link{flowJo_fasinh_trans}} object
}
\examples{
data <- 1:1e3
brks.func <- flowJo_fasinh_breaks()
brks <- brks.func(data)
brks # fasinh space displayed at raw data scale

#transform it to verify it is equal-spaced at transformed scale
trans.func <- flowJo.fasinh()
brks.trans <- trans.func(brks)
brks.trans
}

