% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.filterList}
\alias{fortify.filterList}
\title{Convert a filterList to a data.table useful for ggplot}
\usage{
\method{fortify}{filterList}(model, data, bins = NULL, ...)
}
\arguments{
\item{model}{filterList}

\item{data}{not used}

\item{bins}{used for interpolating polygonGates to prevent it from losing shape when truncated by axis limits}

\item{...}{not used.}
}
\description{
It tries to merge with pData
}

