% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.flowSet}
\alias{fortify}
\alias{fortify.flowSet}
\title{Convert a flowSet to a ggplot-compatible data.frame}
\usage{
\method{fortify}{flowSet}(model, data, ...)
}
\arguments{
\item{model}{flowSet}

\item{data}{not used.}

\item{...}{not used.}
}
\description{
It invokes as.data.frame.flowSet and append the pData
to it so that ggplot can use the pData for facetting.
}

