% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.flowFrame}
\alias{fortify}
\alias{fortify.GatingSet}
\alias{fortify.flowFrame}
\alias{fortify.flowSet}
\title{Convert a flowFrame/flowSet/GatingSet to a ggplot-compatible data.table}
\usage{
\method{fortify}{flowFrame}(model, data, ...)

\method{fortify}{flowSet}(model, data, ...)

\method{fortify}{GatingSet}(model, ...)
}
\arguments{
\item{model}{flowFrame, flowSet or GatingSet}

\item{data}{not used.}

\item{...}{not used.}
}
\value{
data.table

data.table

data.table
}
\description{
It extracts events matrices and appends the pData to it so that ggplot can use the pData for facetting.
}
\examples{
\dontrun{
fortify(fr)#fr is a flowFrame
fortify(fs)#fs is a flowSet/ncdfFlowSet
attr(gs, "subset") <- "CD4" #must attach subset information to GatingSet object before foritfying it
fortify(gs)
}
}

