% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.polygonGate}
\alias{fortify.polygonGate}
\title{Convert a polygonGate to a data.table useful for ggplot}
\usage{
\method{fortify}{polygonGate}(model, data, ...)
}
\arguments{
\item{model}{polygonGate}

\item{data}{not used.}

\item{...}{not used.}
}
\description{
It converts the boundaries slot into a data.table
}

