% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.polygonGate}
\alias{fortify.polygonGate}
\title{Convert a polygonGate to a data.table useful for ggplot}
\usage{
\method{fortify}{polygonGate}(model, data, interpolate = TRUE, ...)
}
\arguments{
\item{model}{polygonGate}

\item{data}{not used.}

\item{interpolate}{whether to interpolate polygon with more vertices. Interpolation is mainly
for the purpose of plotting (so that it won't lose its shape from subsetting through 'limits').
But it is not necessary for other purposes like centroid calculation.}

\item{...}{not used.}
}
\description{
It converts the boundaries slot into a data.table
}

