% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.rectangleGate}
\alias{fortify.rectangleGate}
\title{Convert a rectangleGate to a data.frame useful for ggplot}
\usage{
\method{fortify}{rectangleGate}(model, data, ...)
}
\arguments{
\item{model}{rectangleGate}

\item{data}{not used.}

\item{...}{not used.}
}
\description{
For 2d rectangelGate, it is converted to a geom_polygon format
for 1d, uses geom_vline/hline format.
}

