% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fortify_fs.R
\name{fortify_fs}
\alias{fortify_fs}
\title{Fortify a model into flowSet object}
\usage{
fortify_fs(model, data, ...)
}
\arguments{
\item{model}{model or other R object to convert to data table}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\description{
Method to convert a generic R object into a flowSet useful for ggcyto
}

