% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stats.R
\name{geom_stats}
\alias{geom_stats}
\title{Popluation statistics layer}
\usage{
geom_stats(gate = NULL, ..., value = NULL, type = "percent",
  data_range = NULL, adjust = 0.5, label.padding = unit(0.05, "lines"),
  label.size = 0)
}
\arguments{
\item{gate}{a 'filterList` or character (represent as a population node in GatingSet)
if not supplied, ggcyto then tries to parse the gate from the first geom_gate layer.}

\item{...}{other arguments passed to geom_label layer}

\item{value}{the pre-calculated stats value. when supplied, the stats computing is skipped.}

\item{type}{can be "percent", "count" or "MFI".}

\item{data_range}{the data range for each channels}

\item{adjust}{adjust the position of the centroid. from 0 to 1.}

\item{label.padding, label.size}{arguments passed to geom_label layer}
}
\value{
a geom_popStats layer
}
\description{
It is a virtual layer and will be instanatiated as geom_label layer within ggycto.+ operator.
}
\details{
So it is dedicated for ggcyto context and thus can't not be added to ggplot object directly.
}
\examples{
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
p <- ggcyto(gs, aes(x = CD4, y = CD8), subset = "CD3+") + geom_hex(bins = 64)
p
# add gate and stats layer
p + geom_gate("CD4") + geom_stats()
}

