% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getFlowFrame.R
\name{getFlowFrame}
\alias{getFlowFrame}
\alias{getFlowFrame.GatingHierarchy}
\alias{getFlowFrame.GatingSet}
\alias{getFlowFrame.flowSet}
\title{extract flowFrame data structure from the given R object}
\usage{
getFlowFrame(x)

\method{getFlowFrame}{flowSet}(x)

\method{getFlowFrame}{GatingSet}(x)

\method{getFlowFrame}{GatingHierarchy}(x)
}
\arguments{
\item{x}{flowSet or GatingSet/GatingHierarchy}
}
\value{
an empty flowFrame
}
\description{
Mainly to get the channel and marker information.
}

