% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto.R
\name{+.ggcyto}
\alias{+.ggcyto}
\title{overloaded '+' method for ggcyto}
\usage{
\method{+}{ggcyto}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{ggcyto}}

\item{e2}{A component to add to \code{e1}}
}
\description{
It tries to copy pData from ggcyto object to the gate layers
so that the gate layer does not need to have `pd` to be supplied explicitly by users.
}

