% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-builder.R
\name{+.ggcyto}
\alias{+.ggcyto}
\title{overloaded '+' method for ggcyto}
\usage{
\method{+}{ggcyto}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{ggcyto}}

\item{e2}{A component to add to \code{e1}}
}
\description{
It tries to pass the copy of pData of fs(supplied in ggplot object) to gate objects
so that the gate layer does not need explicit `pd` to be supplied by users.
}

