\name{ggcyto-package}
\alias{ggcyto-package}
\alias{ggcyto}
\docType{package}
\title{
\packageTitle{ggcyto}
}
\description{
\packageDescription{ggcyto}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ggcyto}
\packageIndices{ggcyto}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ggcyto}

Maintainer: \packageMaintainer{ggcyto}
}
\references{
~~ Literature or other references for background information ~~
}


\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
