% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-builder.R
\name{ggcyto.flowSet}
\alias{autoplot}
\alias{autoplot.flowFrame}
\alias{ggcyto.flowSet}
\title{Plot fluorescence intensity in one or two dimension.}
\usage{
\method{ggcyto}{flowSet}(data, mapping, ...)
}
\arguments{
\item{mapping}{passed to ggplot, which defines the dimension of the plot}

\item{...}{other arguments passed to ggplot}

\item{object}{flowFrame or flowSet object}

\item{plotType}{either "histogram" or "density", determines which plot to be displayed}

\item{margin}{whether to enable marginal/bounary events filtering. Default is TRUE.}
}
\value{
a ggplot object
}
\description{
Overloaded autoplot for the cytomertry data structure: flowFrame or flowSet.
It plots the cytometry data with geom_histogram, geom_density or geom_hex.
}
\examples{
data(GvHD)
fs <- GvHD[subset(pData(GvHD), Patient \%in\%5:7 & Visit \%in\% c(5:6))[["name"]]]

#histogram for raw flow data
autoplot(fs, aes(x = `FL1-H`))
# add transformation
autoplot(fs, aes(x = `FL1-H`)) + scale_x_log10()

# disable marginal events filtering
autoplot(fs, aes(x = `FL1-H`), margin = F) + scale_x_log10()

# density
autoplot(fs, aes(x = `FL1-H`), plotType = "density") + scale_x_log10()

# customize border colors
ggplot(fs, aes(x = `FL1-H`)) + facet_wrap(~name) + geom_histogram(colour = "white") + scale_x_log10()

# change the bin width
ggplot(fs, aes(x = `FL1-H`)) + facet_wrap(~name, scale = "free") + geom_histogram(colour = "white", binwidth = 1/10) + scale_x_log10()

#2d plot: default geom_hex plot
autoplot(fs, aes(x = `FSC-H`, y =  `SSC-H`))

# add contour
autoplot(fs, aes(x = `FSC-H`, y =  `SSC-H`)) + geom_density2d(colour = "black")

# change the faceting
autoplot(fs, aes(x = `FSC-H`, y =  `SSC-H`)) + facet_grid(Patient~Visit)
}

