% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcyto_flowSet.R
\name{ggcyto.flowSet}
\alias{ggcyto.flowSet}
\alias{ggcyto.ncdfFlowList}
\title{Create a new ggcyto plot from a flowSet}
\usage{
\method{ggcyto}{flowSet}(data, mapping, filter = NULL,
  max_nrow_to_plot = 50000, ...)

\method{ggcyto}{ncdfFlowList}(data, ...)
}
\arguments{
\item{data}{default flowSet for plot}

\item{mapping}{default list of aesthetic mappings (these can be colour,
size, shape, line type -- see individual geom functions for more details)}

\item{filter}{a flowcore gate object or a function that takes flowSet and channels as input and returns a data-dependent flowcore gate
The gate is used to filter the flow data before it is plotted.}

\item{max_nrow_to_plot}{the maximum number of cells to be plotted. When the actual data exceeds it, The subsampling process will be triggered to speed up plotting. Default is 2e5.To turn off the subsampling, simply set it to a large enough number or Inf}

\item{...}{ignored}
}
\value{
a ggcyto_GatingSet object which is a subclass of ggcyto class.
}
\description{
Create a new ggcyto plot from a flowSet
}
\examples{

data(GvHD)
fs <- GvHD[subset(pData(GvHD), Patient \%in\%5:7 & Visit \%in\% c(5:6))[["name"]]]
# 1d histogram/densityplot
p <- ggcyto(fs, aes(x = `FSC-H`)) 
#facet_wrap(~name)` is used automatically
p1 <- p + geom_histogram() 
p1
#overwriting the default faceeting
p1 + facet_grid(Patient~Visit)

#display density
p + geom_density()

#you can use ggjoy package to display stacked density plot
require(ggjoy)
#stack by fcs file ('name')
p + geom_joy(aes(y = name)) + facet_null() #facet_null is used to remove the default facet_wrap (by 'name' column)
#or to stack by Visit and facet by patient
p + geom_joy(aes(y = Visit)) + facet_grid(~Patient)

# 2d scatter/dot plot
p <- ggcyto(fs, aes(x = `FSC-H`, y =  `SSC-H`))
p <- p + geom_hex(bins = 128)
p

}
