% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto-GatingSet.R
\name{ggcyto.GatingSet}
\alias{ggcyto.GatingSet}
\title{Create a new ggcyto plot from a flowSet}
\usage{
\method{ggcyto}{GatingSet}(data, mapping, subset = "_parent_", ...)
}
\arguments{
\item{data}{GatingSet to plot}

\item{mapping}{default list of aesthetic mappings (these can be colour,
size, shape, line type -- see individual geom functions for more details)}

\item{subset}{character that specifies the node path or node name in the GatingSet.
Default is "_parent_", which will be substitute with the actual node name
based on the geom_gate layer to be added later.}

\item{...}{ignored}
}
\description{
Create a new ggcyto plot from a flowSet
}

