% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto_GatingSet.R
\name{ggcyto.GatingSet}
\alias{ggcyto.GatingHierarchy}
\alias{ggcyto.GatingSet}
\title{Create a new ggcyto plot from a GatingSet}
\usage{
\method{ggcyto}{GatingSet}(data, mapping, subset = "_parent_", ...)

\method{ggcyto}{GatingHierarchy}(data, ...)
}
\arguments{
\item{data}{GatingSet to plot}

\item{mapping}{default list of aesthetic mappings (these can be colour,
 size, shape, line type -- see individual geom functions for more details)}

\item{subset}{character that specifies the node path or node name in the GatingSet.
Default is "_parent_", which will be substitute with the actual node name
based on the geom_gate layer to be added later.}

\item{...}{ignored}
}
\value{
a ggcyto_GatingSet object which is a subclass of ggcyto_flowSet class.
}
\description{
Create a new ggcyto plot from a GatingSet
}
\examples{
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
# 2d plot
ggcyto(gs, aes(x = CD4, y = CD8), subset = "CD3+") + geom_hex(bins = 64)

# 1d plot
ggcyto(gs, aes(x = CD4), subset = "CD3+")  + geom_density()
}

