% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ggcyto.R
\docType{class}
\name{ggcyto-class}
\alias{ggcyto}
\alias{ggcyto-class}
\alias{ggcyto.default}
\alias{ggcyto_GatingLayout-class}
\alias{ggcyto_GatingSet-class}
\alias{ggcyto_flowSet-class}
\title{Create a new ggcyto plot}
\usage{
ggcyto(data = NULL, ...)

\method{ggcyto}{default}(data = NULL, mapping = aes(), ...)
}
\arguments{
\item{data}{default cytometry data set.(flowSet,flowFrame)}

\item{...}{other arguments passed to specific methods}
}
\description{
\code{ggcyto()} initializes a ggcyto object that inherits ggplot class.
Similarly the + operator can be used to add layers to the
existing ggcyto object.
}
\details{
To invoke \code{ggcyto}:
\itemize{
   \item \code{ggcyto(fs, aes(x, y, <other aesthetics>))}
  }
}
\examples{
\dontrun{
#construct the `ggcyto` object (inherits from `ggplot` class)
p <- ggcyto(fs, aes(x = `FSC-H`))
p + geom_histogram()

# display density/area
p + geom_density()
p + geom_area(stat = "density")

# 2d scatter plot
p <- ggcyto(fs, aes(x = `FSC-H`, y =  `SSC-H`))
p + stat_binhex(bin = 128)
}
}
\keyword{internal}

