% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/ggcyto.R
\docType{class}
\name{ggcyto-class}
\alias{ggcyto-class}
\alias{ggcyto_flowSet-class}
\alias{ggcyto_GatingSet-class}
\alias{ggcyto_GatingLayout-class}
\alias{ggcyto}
\alias{ggcyto.default}
\title{Plot cytometry data using the ggcyto API}
\usage{
ggcyto(data = NULL, ...)

\method{ggcyto}{default}(data = NULL, mapping = aes(), ...)
}
\arguments{
\item{data}{The data source. A core cytometry data structure. (flowSet,flowFrame, GatingSet or GatingHierarchy)}

\item{...}{other arguments passed to specific methods}
}
\description{
\code{ggcyto()} initializes a ggcyto object that inherits ggplot class.
Similarly the + operator can be used to add layers to the
existing ggcyto object.
}
\details{
To invoke \code{ggcyto}:
\itemize{
   \item \code{ggcyto(fs, aes(x, y, <other aesthetics>))}
  }

@return ggcyto object
}
\examples{

data(GvHD)
fs <- GvHD[1:3]
#construct the `ggcyto` object (inherits from `ggplot` class)
p <- ggcyto(fs, aes(x = `FSC-H`)) 
p + geom_histogram() 

# display density/area
p + geom_density()
p + geom_area(stat = "density") 

# 2d scatter plot
p <- ggcyto(fs, aes(x = `FSC-H`, y =  `SSC-H`))
p + geom_hex(bins = 128)
# do it programatically through aes_string and variables
col1 <- "`FSC-H`" #note that the dimension names with special characters needs to be quoted by backticks
col2 <- "`SSC-H`"
ggcyto(fs, aes_string(col1,col2)) + geom_hex()
}
\keyword{internal}
