% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto_GatingSet.R
\docType{methods}
\name{+.ggcyto_GatingSet}
\alias{+,ggcyto_GatingSet,ANY-method}
\alias{+.ggcyto_GatingSet}
\title{overloaded '+' method for ggcyto.gs}
\usage{
\method{+}{ggcyto_GatingSet}(e1, e2)

\S4method{+}{ggcyto_GatingSet,ANY}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{ggcyto}}

\item{e2}{A component to add to \code{e1}}
}
\value{
ggcyto_GatingSet object
}
\description{
It takes care the speical format of some ggcyto layers. For example geom_gate or geom_stats layer with just gate(population) name specified,
It only supports some special axis transformations. (See examples below)
}
\examples{
dataDir <- system.file("extdata",package="flowWorkspaceData")
 gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
 p <- ggcyto(gs, aes(x = CD4, y = CD8), subset = "CD3+") + geom_hex(bins = 64)
 p <- p + geom_gate("CD4") + geom_stats() #plot CD4 gate and it is stats
 p
 p + axis_x_inverse_trans() #inverse transform the x axis into raw scale
}

