% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto_flowSet.R
\name{is.ggcyto_flowSet}
\alias{is.ggcyto_flowSet}
\title{Reports whether x is a ggcyto_flowSet object}
\usage{
is.ggcyto_flowSet(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE or FALSE
}
\description{
Reports whether x is a ggcyto_flowSet object
}
\examples{
data(GvHD)
fs <- GvHD[1:2]
p <- ggcyto(fs, aes(x = `FSC-H`))
is.ggcyto_flowSet(p)
}

