% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcyto_par.R
\name{is.ggcyto_par}
\alias{is.ggcyto_par}
\title{Reports whether x is a ggcyto_par object}
\usage{
is.ggcyto_par(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE or FALSE
}
\description{
Reports whether x is a ggcyto_par object
}
\examples{
myPar <- ggcyto_par_set(limits = "instrument")
is.ggcyto_par(myPar)
}
