% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geom_hvline.R
\name{my_compute_aesthetics}
\alias{my_compute_aesthetics}
\title{override the original Layer$compute_aesthetics method so that it could make it for the layer
that contains 1d data but uses the 2d aes}
\usage{
my_compute_aesthetics(., data, plot)
}
\description{
This is the hook invoked by ggplot_build at early stage of plotting.
We need to match the 1d gate definition with the mapping
stored in the original aes (from 'plot$mapping') so that the line can be determined as either horizontal or vetical
}
\details{
And so far this is the best place where we can look up the original mapping and store the axis information.
Idealy it could have been be done at `map_statistic` step. Unfortunetly, the data would have already been generalized at that point and
thus it won't be able to do the matching otherwise.

It would be nice if we could also hack to associate pData from plot$data here. Unforutntely the data
is already facetted before this step and there is no way to access the facetted plot$data in order to
reassign the 'panel' info for layer data.
}

