% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto_GatingLayout.R
\docType{methods}
\name{+.ggcyto_GatingLayout}
\alias{+,ggcyto_GatingLayout,ANY-method}
\alias{+.ggcyto_GatingLayout}
\title{overloaded '+' method for ggcyto_gate_layout}
\usage{
\method{+}{ggcyto_GatingLayout}(e1, e2)

\S4method{+}{ggcyto_GatingLayout,ANY}(e1, e2)
}
\arguments{
\item{e1}{ggcyto_gate_layout}

\item{e2}{any ggplot layer}
}
\value{
a modified ggcyto_gate_layout object

a GatingLayout object
}
\description{
It adds the layer specified by 'e2' to each individual ggplot object stored in ggcyto_gate_layout
}
\examples{
#autplot for GatingSet
dataDir <- system.file("extdata",package="flowWorkspaceData")
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
gh <- gs[[1]]
p <- autoplot(gh)
class(p)
# customize the font size of strip text for each ggcyo plots contained in GatingLayout object
p + theme(strip.text = element_text(size = 14))
}

