% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto.R
\name{print.ggcyto}
\alias{plot.ggcyto}
\alias{print.ggcyto}
\title{Draw ggcyto on current graphics device.}
\usage{
\method{print}{ggcyto}(x, ...)

\method{plot}{ggcyto}(x, ...)
}
\arguments{
\item{x}{ggcyto object to display}

\item{...}{other arguments not used by this method}
}
\description{
A wrapper for print.ggplot.
}

