% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcyto.R
\name{print.ggcyto}
\alias{print.ggcyto}
\alias{print,ggcyto-method}
\alias{plot.ggcyto}
\alias{show.ggcyto}
\alias{show,ggcyto-method}
\title{Draw ggcyto on current graphics device.}
\usage{
\method{print}{ggcyto}(x, ...)

\method{plot}{ggcyto}(x, ...)

\method{show}{ggcyto}(object)
}
\arguments{
\item{x}{ggcyto object to display}

\item{...}{other arguments not used by this method}

\item{object}{ggcyto object}
}
\value{
nothing
}
\description{
A wrapper for print.ggplot. It converts the ggcyto to conventional ggplot object before printing it.
This is usually invoked automatically when a ggcyto object is returned to R console.
}
