% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggcyto_GatingLayout.R
\docType{methods}
\name{print.ggcyto_GatingLayout}
\alias{print.ggcyto_GatingLayout}
\alias{show,ggcyto_GatingLayout-method}
\alias{show.ggcyto_GatingLayout}
\title{print method for ggcyto_gate_layout class
It calls arrangeGrob to arrange a list of ggplot objects stored as ggcyto_gate_layout object}
\usage{
\method{print}{ggcyto_GatingLayout}(x, ...)

\method{show}{ggcyto_GatingLayout}(object)

\S4method{show}{ggcyto_GatingLayout}(object)
}
\arguments{
\item{x}{ggcyto_gate_layout, which is essentially a list of ggplot objects}

\item{...}{other arguments passed to arrangeGrob}

\item{object}{ggcyto_GatingLayout}
}
\value{
nothing
}
\description{
print method for ggcyto_gate_layout class
It calls arrangeGrob to arrange a list of ggplot objects stored as ggcyto_gate_layout object
}

