% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_biexp.R
\name{scale_x_flowJo_biexp}
\alias{scale_x_flowJo_biexp}
\alias{scale_y_flowJo_biexp}
\title{flowJo biexponential scale}
\usage{
scale_x_flowJo_biexp(..., maxValue = 262144, widthBasis = -10, pos = 4.5,
  neg = 0, pretty = FALSE)

scale_y_flowJo_biexp(..., maxValue = 262144, widthBasis = -10, pos = 4.5,
  neg = 0, pretty = FALSE)
}
\arguments{
\item{...}{common continuous scale parameters passed to 'continuous_scale' (not used currently)}

\item{maxValue,widthBasis,pos,neg}{see 'help(flowJoTrans')}

\item{pretty}{whether to display the breaks in pretty format}
}
\value{
ScaleContinuous object
}
\description{
flowJo biexponential scale
}
\examples{
data(GvHD)
fr <- GvHD[[1]]
p <- ggcyto(fr, aes(x = `FL1-H`)) + geom_density()
#display at raw scale
p
#display at transformed scale
p + scale_x_flowJo_biexp(maxValue = 1e4, widthBasis = 0)
}

