% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scales_flowJo_fasinh.R
\name{scale_x_flowJo_fasinh}
\alias{scale_x_flowJo_fasinh}
\alias{scale_y_flowJo_fasinh}
\title{flowJo inverse hyperbolic sine scale}
\usage{
scale_x_flowJo_fasinh(..., m = 4, t = 1200)

scale_y_flowJo_fasinh(..., m = 4, t = 1200)
}
\arguments{
\item{...}{common continuous scale parameters passed to 'continuous_scale' (not used currently)}

\item{m,t}{see 'help(flowJo.fasinh')}
}
\description{
flowJo inverse hyperbolic sine scale
}

