% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_position.R
\name{stat_position}
\alias{stat_position}
\alias{stat_position.filter}
\alias{stat_position.filterList}
\alias{stat_position.list}
\title{Compute the positions of the population statistics based on the geometic gate centroid for a ggcyto plot.}
\usage{
stat_position(gate, ...)

\method{stat_position}{filter}(gate, ...)

\method{stat_position}{filterList}(gate, ...)

\method{stat_position}{list}(gate, ...)
}
\arguments{
\item{gate}{a flowCore filter}

\item{...}{other arguments
adjust adjust the position of the centroid

abs logical

data_range the actual data range}
}
\value{
a data.table

the gate centroid coordinates
}
\description{
It is usually not called directly by user but mainly used by compute_stats function (which is called by ggcyto add method when geom_states layer is added).
}
\examples{
data(GvHD)
fs <- GvHD[1:4]
rect.g <- rectangleGate(list("FSC-H" =  c(300,500), "SSC-H" = c(50,200)))
rect.gates <- sapply(sampleNames(fs), function(sn)rect.g)
stat_position(rect.gates)
}
