% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_gate.R
\name{transform-gate}
\alias{transform-gate}
\alias{transform}
\alias{transform,filter-method}
\alias{transform-filterList-method}
\alias{rescale_gate}
\alias{rescale_gate.polygonGate}
\alias{rescale_gate.ellipsoidGate}
\alias{rescale_gate.quadGate}
\alias{rescale_gate.rectangleGate}
\title{rescale methods for gates}
\usage{
transform(`_data`, ...)

rescale_gate(gate, trans, param)
}
\arguments{
\item{_data}{the filter or filterList object. Currently support polygonGate, ellipsoidGate, rectangleGate and quadGate.}

\item{...}{trans the transformation function or transformList object
param the parameter/dimension to be transformed. When trans is transformList object, param is not needed since it is derived from transformList.}

\item{gate}{gate object}

\item{trans}{the transformation function}

\item{param}{the parameter/dimension to be transformed.}
}
\value{
the transformed filter/filterList object
}
\description{
rescale the gate coordinates with the transformation provided
}
