% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_gate.R
\docType{methods}
\name{transform,filter-method}
\alias{transform,filter-method}
\alias{transform,filterList-method}
\alias{transform,list-method}
\alias{transform_gate}
\alias{transform_gate.polygonGate}
\alias{transform_gate.ellipsoidGate}
\alias{transform_gate.rectangleGate}
\alias{transform_gate.quadGate}
\title{transform methods for gates}
\usage{
\S4method{transform}{filter}(`_data`, ...)

\S4method{transform}{filterList}(`_data`, ...)

\S4method{transform}{list}(`_data`, ...)

transform_gate(`_data`, ...)

\method{transform_gate}{polygonGate}(gate, trans, param)

\method{transform_gate}{ellipsoidGate}(gate, ...)

\method{transform_gate}{rectangleGate}(gate, trans, param)

\method{transform_gate}{quadGate}(gate, trans, param)
}
\arguments{
\item{_data}{the filter or filterList object. Currently support polygonGate, ellipsoidGate, rectangleGate and quadGate.}

\item{...}{trans the transformation function or transformList object
param the parameter/dimension to be transformed. When trans is transformList object, param is not needed since it is derived from transformList.}

\item{gate}{gate object}

\item{trans}{the transformation function}

\item{param}{the parameter/dimension to be transformed.}

\item{_data}{the gate object. Currently support polygonGate, ellipsoidGate, rectangleGate and quadGate.}
}
\value{
the transformed filter/filterList object

the transformed gate object
}
\description{
transform methods for gates

transform methods for gates
}
